add js
шаблонизация
