+++ index.html
@@ -173,7 +173,7 @@
       <button class="error__button">Попробовать снова</button>
     </div>
   </template>
-  <script src="../js_project_9/js/main.js"></script>
+  <script src="./js/main.js" type="module"></script>
 </body>
 
 </html>
