+++ js/helpers.js
@@ -1,0 +1,4 @@
+export function randomInteger(min, max) {
+    let rand = min - 0.5 + Math.random() * (max - min + 1);
+    return Math.round(rand);
+}
